refactor(Popularity): move types from derivedTypes.ts => rezervoTypes.ts
