--- types/derivedTypes.ts
@@ -1,10 +1,0 @@
-export enum ClassPopularity {
-    Unknown = "Denne timen gikk ikke forrige uke, og har derfor ukjent popularitet.",
-    Low = "Denne timen har vanligvis mange ledige plasser.",
-    Medium = "Denne timen har vanligvis noen ledige plasser.",
-    High = "Denne timen er vanligvis full.",
-}
-
-export type ClassPopularityIndex = {
-    [recurrentId: string]: ClassPopularity;
-};
